// src/Types/DsTypographyVariants.ts
export interface DsTypographyVariants {
  displayBoldLarge: React.CSSProperties
  displayBoldMedium: React.CSSProperties
  displayBoldSmall: React.CSSProperties
  displayBoldItalicLarge: React.CSSProperties
  displayBoldItalicMedium: React.CSSProperties
  displayBoldItalicSmall: React.CSSProperties
  headingBoldExtraLarge: React.CSSProperties
  headingBoldLarge: React.CSSProperties
  headingBoldMedium: React.CSSProperties
  headingBoldSmall: React.CSSProperties
  headingBoldExtraSmall: React.CSSProperties
  subheadingSemiboldLarge: React.CSSProperties
  subheadingSemiboldDefault: React.CSSProperties
  headingBoldItalicExtraLarge: React.CSSProperties
  headingBoldItalicLarge: React.CSSProperties
  headingBoldItalicMedium: React.CSSProperties
  headingBoldItalicSmall: React.CSSProperties
  headingBoldItalicExtraSmall: React.CSSProperties
  bodyRegularLarge: React.CSSProperties
  bodyRegularMedium: React.CSSProperties
  bodyRegularSmall: React.CSSProperties
  bodyBoldLarge: React.CSSProperties
  bodyBoldMedium: React.CSSProperties
  bodyBoldSmall: React.CSSProperties
  supportRegularInfo: React.CSSProperties
  supportRegularFootnote: React.CSSProperties
  supportBoldTextButton: React.CSSProperties
  supportRegularMetadata: React.CSSProperties
}

export interface DsTypographyPropsVariantOverrides {
  displayBoldLarge: true
  displayBoldMedium: true
  displayBoldSmall: true
  displayBoldItalicLarge: true
  displayBoldItalicMedium: true
  displayBoldItalicSmall: true
  headingBoldExtraLarge: true
  headingBoldLarge: true
  headingBoldMedium: true
  headingBoldSmall: true
  headingBoldExtraSmall: true
  subheadingSemiboldLarge: true
  subheadingSemiboldDefault: true
  headingBoldItalicExtraLarge: true
  headingBoldItalicLarge: true
  headingBoldItalicMedium: true
  headingBoldItalicSmall: true
  headingBoldItalicExtraSmall: true
  bodyRegularLarge: true
  bodyRegularMedium: true
  bodyRegularSmall: true
  bodyBoldLarge: true
  bodyBoldMedium: true
  bodyBoldSmall: true
  supportRegularInfo: true
  supportRegularFootnote: true
  supportBoldTextButton: true
  supportRegularMetadata: true
}